+++ delta/src/common/error.rs
