+++ delta/src/neuralnet/layers/conv_2d.rs
@@ -30,7 +30,7 @@
 use crate::activations::Activation;
 use crate::common::shape::Shape;
 use crate::common::tensor_ops::Tensor;
-use crate::common::LayerError;
+use crate::neuralnet::layers::error::LayerError;
 use crate::neuralnet::layers::Layer;
 
 /// A struct representing a 2D convolutional layer.
