+++ delta/src/neuralnet/layers/dense.rs
@@ -27,10 +27,11 @@
 //! OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 //! OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use crate::common::{LayerError, Shape, Tensor};
+use crate::common::{Shape, Tensor};
 use log::debug;
 use serde_json;
 use crate::activations::Activation;
+use crate::neuralnet::layers::error::LayerError;
 use crate::neuralnet::layers::Layer;
 use crate::optimizers::Optimizer;
