+++ delta/src/neuralnet/layers/error.rs
@@ -27,31 +27,39 @@
 //! OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 //! OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+
 use std::fmt;
+use crate::optimizers::error::OptimizerError;
 
-use super::LayerError;
-
+/// Errors that can occur in the Dense layer.
 #[derive(Debug)]
-pub enum ModelError {
-    MissingOptimizer,
-    MissingLossFunction,
-    DatasetError(String),
-    TrainingError(String),
-    LayerError(LayerError),
+pub enum LayerError {
+    /// Error when weights are not initialized.
+    UninitializedWeights,
+    /// Error when bias is not initialized.
+    UninitializedBias,
+    /// Error when input is not set for backward pass.
+    UninitializedInput,
+    /// Error related to invalid shape.
+    InvalidShape,
+    /// Error when input is not set.
+    MissingInput,
+
+    /// Error when an optimizer error occurs.
+    OptimizerError(OptimizerError),
 }
 
-impl fmt::Display for ModelError {
+impl fmt::Display for LayerError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            ModelError::MissingOptimizer => write!(f, "Optimizer must be set before training"),
-            ModelError::MissingLossFunction => {
-                write!(f, "Loss function must be set before training")
-            }
-            ModelError::DatasetError(msg) => write!(f, "Dataset error: {}", msg),
-            ModelError::TrainingError(msg) => write!(f, "Training error: {}", msg),
-            ModelError::LayerError(err) => write!(f, "Layer error: {}", err),
+            LayerError::UninitializedWeights => write!(f, "Weights must be initialized"),
+            LayerError::UninitializedBias => write!(f, "Bias must be initialized"),
+            LayerError::UninitializedInput => write!(f, "Input must be initialized"),
+            LayerError::InvalidShape => write!(f, "Invalid shape"),
+            LayerError::MissingInput => write!(f, "Input must be set"),
+            LayerError::OptimizerError(err) => write!(f, "Optimizer error: {}", err),
         }
     }
 }
 
-impl std::error::Error for ModelError {}
+impl std::error::Error for LayerError {}
