+++ delta/src/neuralnet/models/sequential.rs
@@ -33,10 +33,11 @@ use std::io::Write;
 use std::path::Path;
 use std::time::Instant;
 
-use crate::common::{ModelError, Tensor};
+use crate::common::Tensor;
 use crate::dataset::{Dataset, ImageDatasetOps};
 use crate::losses::Loss;
 use crate::neuralnet::layers::Layer;
+use crate::neuralnet::models::error::ModelError;
 use crate::optimizers::Optimizer;
 
 /// A sequential model that contains a list of layers, an optimizer, and a loss function.
