+++ delta/src/optimizers/error.rs
